layermodal refaktorointi -> layeritem eriytys, focus sailyy
